Move folder constants to folder app
