+++ practice/django_practice/departments/models.py
@@ -1,4 +1,3 @@
-import uuid
 from django.db import models
 
 from django_practice.models import BaseModel
@@ -14,9 +13,6 @@ class Department(BaseModel):
         description (TextField): The description of the department
     """
 
-    id = models.UUIDField(
-        primary_key=True, default=uuid.uuid4, editable=False
-    )
     name = models.CharField(max_length=100, unique=True)
     description = models.TextField(max_length=200)
