+++ practice/django_practice/django_practice/constants/enum.py
