+++ practice/django_practice/django_practice/models.py
@@ -1,3 +1,4 @@
+import uuid
 from django.db import models
 
 
@@ -8,6 +9,7 @@ class BaseModel(models.Model):
     """
 
     create_at = models.DateTimeField(auto_now_add=True)
+    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
 
     class Meta:
         abstract = True
